Extract out TryRestore so project and tool restores can be isolated
