+++ src/NuGet.Core/NuGet.Commands/RestoreCommand/RestoreCommand.cs
@@ -8,6 +8,7 @@ using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
@@ -257,14 +258,22 @@ namespace NuGet.Commands
                 var allInstalledPackages = new HashSet<LibraryIdentity>();
                 var contextForTool = CreateRemoteWalkContext(_request);
                 var walker = new RemoteDependencyWalker(contextForTool);
-                var result = await TryRestore(tool.LibraryRange,
-                                              projectFrameworkRuntimePairs,
-                                              allInstalledPackages,
-                                              localRepository,
-                                              walker,
-                                              contextForTool,
-                                              writeToLockFile: true,
-                                              token: token);
+                var projectRestoreRequest = new ProjectRestoreRequest(
+                    _request,
+                    toolPackageSpec,
+                    null,
+                    new Dictionary<NuGetFramework, RuntimeGraph>(), 
+                    _runtimeGraphCacheByPackage);
+                var projectRestoreCommand = new ProjectRestoreCommand(_logger, projectRestoreRequest);
+                var result = await projectRestoreCommand.TryRestore(
+                    tool.LibraryRange,
+                    projectFrameworkRuntimePairs,
+                    allInstalledPackages,
+                    localRepository,
+                    walker,
+                    contextForTool,
+                    writeToLockFile: true,
+                    token: token);
 
                 var graphs = result.Item2;
                 if (!result.Item1)
@@ -397,14 +406,22 @@ namespace NuGet.Commands
             var runtimeIds = RequestRuntimeUtility.GetRestoreRuntimes(_request);
             var projectFrameworkRuntimePairs = CreateFrameworkRuntimePairs(_request.Project, runtimeIds);
 
-            var result = await TryRestore(projectRange,
-                                          projectFrameworkRuntimePairs,
-                                          allInstalledPackages,
-                                          localRepository,
-                                          remoteWalker,
-                                          context,
-                                          writeToLockFile: true,
-                                          token: token);
+            var projectRestoreRequest = new ProjectRestoreRequest(
+                _request,
+                _request.Project,
+                _request.ExistingLockFile,
+                _runtimeGraphCache,
+                _runtimeGraphCacheByPackage);
+            var projectRestoreCommand = new ProjectRestoreCommand(_logger, projectRestoreRequest);
+            var result = await projectRestoreCommand.TryRestore(
+                projectRange,
+                projectFrameworkRuntimePairs,
+                allInstalledPackages,
+                localRepository,
+                remoteWalker,
+                context,
+                writeToLockFile: true,
+                token: token);
 
             var success = result.Item1;
             var runtimes = result.Item3;
@@ -439,7 +456,7 @@ namespace NuGet.Commands
             // Walk additional runtime graphs for supports checks
             if (_success && _request.CompatibilityProfiles.Any())
             {
-                var compatibilityResult = await TryRestore(projectRange,
+                var compatibilityResult = await projectRestoreCommand.TryRestore(projectRange,
                                                           _request.CompatibilityProfiles,
                                                           allInstalledPackages,
                                                           localRepository,
@@ -511,132 +528,6 @@ namespace NuGet.Commands
             }
 
             return context;
-        }
-
-        private async Task<Tuple<bool, List<RestoreTargetGraph>, RuntimeGraph>> TryRestore(LibraryRange projectRange,
-            IEnumerable<FrameworkRuntimePair> frameworkRuntimePairs,
-            HashSet<LibraryIdentity> allInstalledPackages,
-            NuGetv3LocalRepository localRepository,
-            RemoteDependencyWalker remoteWalker,
-            RemoteWalkContext context,
-            bool writeToLockFile,
-            CancellationToken token)
-        {
-            var allRuntimes = RuntimeGraph.Empty;
-            var frameworkTasks = new List<Task<RestoreTargetGraph>>();
-            var graphs = new List<RestoreTargetGraph>();
-            var runtimesByFramework = frameworkRuntimePairs.ToLookup(p => p.Framework, p => p.RuntimeIdentifier);
-
-            foreach (var pair in runtimesByFramework)
-            {
-                _logger.LogVerbose(string.Format(CultureInfo.CurrentCulture, Strings.Log_RestoringPackages, pair.Key.DotNetFrameworkName));
-
-                frameworkTasks.Add(WalkDependenciesAsync(projectRange,
-                    pair.Key,
-                    remoteWalker,
-                    context,
-                    writeToLockFile: writeToLockFile,
-                    token: token));
-            }
-
-            var frameworkGraphs = await Task.WhenAll(frameworkTasks);
-
-            graphs.AddRange(frameworkGraphs);
-
-            if (!ResolutionSucceeded(frameworkGraphs))
-            {
-                return Tuple.Create(false, graphs, allRuntimes);
-            }
-
-            await InstallPackagesAsync(graphs,
-                    _request.PackagesDirectory,
-                    allInstalledPackages,
-                    _request.MaxDegreeOfConcurrency,
-                    token);
-
-            // Clear the in-memory cache for newly installed packages
-            localRepository.ClearCacheForIds(allInstalledPackages.Select(package => package.Name));
-
-            // Resolve runtime dependencies
-            var runtimeGraphs = new List<RestoreTargetGraph>();
-            if (runtimesByFramework.Count > 0)
-            {
-                var runtimeTasks = new List<Task<RestoreTargetGraph[]>>();
-                foreach (var graph in graphs)
-                {
-                    // Get the runtime graph for this specific tfm graph
-                    var runtimeGraph = GetRuntimeGraph(graph, localRepository);
-                    var runtimeIds = runtimesByFramework[graph.Framework];
-
-                    // Merge all runtimes for the output
-                    allRuntimes = RuntimeGraph.Merge(allRuntimes, runtimeGraph);
-
-                    runtimeTasks.Add(WalkRuntimeDependenciesAsync(projectRange,
-                        graph,
-                        runtimeIds.Where(rid => !string.IsNullOrEmpty(rid)),
-                        remoteWalker,
-                        context,
-                        localRepository,
-                        runtimeGraph,
-                        writeToLockFile: writeToLockFile,
-                        token: token));
-                }
-
-                foreach (var runtimeSpecificGraph in (await Task.WhenAll(runtimeTasks)).SelectMany(g => g))
-                {
-                    runtimeGraphs.Add(runtimeSpecificGraph);
-                }
-
-                graphs.AddRange(runtimeGraphs);
-
-                if (!ResolutionSucceeded(runtimeGraphs))
-                {
-                    return Tuple.Create(false, graphs, allRuntimes);
-                }
-
-                // Install runtime-specific packages
-                await InstallPackagesAsync(runtimeGraphs,
-                    _request.PackagesDirectory,
-                    allInstalledPackages,
-                    _request.MaxDegreeOfConcurrency,
-                    token);
-
-                // Clear the in-memory cache for newly installed packages
-                localRepository.ClearCacheForIds(allInstalledPackages.Select(package => package.Name));
-            }
-
-            return Tuple.Create(true, graphs, allRuntimes);
-        }
-
-        private bool ResolutionSucceeded(IEnumerable<RestoreTargetGraph> graphs)
-        {
-            var success = true;
-            foreach (var graph in graphs)
-            {
-                if (graph.Conflicts.Any())
-                {
-                    success = false;
-                    _logger.LogError(string.Format(CultureInfo.CurrentCulture, Strings.Log_FailedToResolveConflicts, graph.Name));
-                    foreach (var conflict in graph.Conflicts)
-                    {
-                        _logger.LogError(string.Format(CultureInfo.CurrentCulture, Strings.Log_ResolverConflict, 
-                            conflict.Name,
-                            string.Join(", ", conflict.Requests)));
-                    }
-                }
-                if (graph.Unresolved.Any())
-                {
-                    success = false;
-                    foreach (var unresolved in graph.Unresolved)
-                    {
-                        _logger.LogError(string.Format(CultureInfo.CurrentCulture, Strings.Log_UnresolvedDependency, unresolved.Name,
-                            unresolved.VersionRange.ToNonSnapshotRange().PrettyPrint(),
-                            graph.Name));
-                    }
-                }
-            }
-
-            return success;
         }
 
         private MSBuildRestoreResult RestoreMSBuildFiles(PackageSpec project,
@@ -760,272 +651,6 @@ namespace NuGet.Commands
                 // Null out all target types, these did not exist in v1
                 library.Type = null;
             }
-        }
-        
-
-        private Task<RestoreTargetGraph> WalkDependenciesAsync(LibraryRange projectRange,
-            NuGetFramework framework,
-            RemoteDependencyWalker walker,
-            RemoteWalkContext context,
-            bool writeToLockFile,
-            CancellationToken token)
-        {
-            return WalkDependenciesAsync(projectRange,
-                framework,
-                runtimeIdentifier: null,
-                runtimeGraph: RuntimeGraph.Empty,
-                walker: walker,
-                context: context,
-                writeToLockFile: writeToLockFile,
-                token: token);
-        }
-
-        private async Task<RestoreTargetGraph> WalkDependenciesAsync(LibraryRange projectRange,
-            NuGetFramework framework,
-            string runtimeIdentifier,
-            RuntimeGraph runtimeGraph,
-            RemoteDependencyWalker walker,
-            RemoteWalkContext context,
-            bool writeToLockFile,
-            CancellationToken token)
-        {
-            var name = FrameworkRuntimePair.GetName(framework, runtimeIdentifier);
-            var graphs = new List<GraphNode<RemoteResolveResult>>();
-            if (_request.ExistingLockFile != null && _request.ExistingLockFile.IsLocked)
-            {
-                // Walk all the items in the lock file target and just synthesize the outer graph
-                var target = _request.ExistingLockFile.GetTarget(framework, runtimeIdentifier);
-
-                token.ThrowIfCancellationRequested();
-                if (target != null)
-                {
-                    foreach (var targetLibrary in target.Libraries)
-                    {
-                        token.ThrowIfCancellationRequested();
-
-                        var library = _request.ExistingLockFile.GetLibrary(targetLibrary.Name, targetLibrary.Version);
-                        if (library == null)
-                        {
-                            _logger.LogWarning(string.Format(CultureInfo.CurrentCulture, Strings.Log_LockFileMissingLibraryForTargetLibrary, 
-                                targetLibrary.Name,
-                                targetLibrary.Version,
-                                target.Name));
-                            continue; // This library is not in the main lockfile?
-                        }
-
-                        var range = new LibraryRange()
-                        {
-                            Name = library.Name,
-                            TypeConstraint = LibraryDependencyTargetUtils.Parse(library.Type),
-                            VersionRange = new VersionRange(
-                                minVersion: library.Version,
-                                includeMinVersion: true,
-                                maxVersion: library.Version,
-                                includeMaxVersion: true)
-                        };
-                        graphs.Add(await walker.WalkAsync(
-                            range,
-                            framework,
-                            runtimeIdentifier,
-                            runtimeGraph,
-                            recursive: false));
-                    }
-                }
-            }
-            else
-            {
-                graphs.Add(await walker.WalkAsync(
-                    projectRange,
-                    framework,
-                    runtimeIdentifier,
-                    runtimeGraph,
-                    recursive: true));
-            }
-
-            // Resolve conflicts
-            _logger.LogVerbose(string.Format(CultureInfo.CurrentCulture, Strings.Log_ResolvingConflicts, name));
-
-            // Flatten and create the RestoreTargetGraph to hold the packages
-            var result = RestoreTargetGraph.Create(writeToLockFile, runtimeGraph, graphs, context, _logger, framework, runtimeIdentifier);
-
-            // Check if the dependencies got bumped up
-            // ...but not if there is an existing locked lock file.
-            if (_request.ExistingLockFile == null || !_request.ExistingLockFile.IsLocked)
-            {
-                // No lock file, OR the lock file is unlocked, so check dependencies
-                CheckDependencies(result, _request.Project.Dependencies);
-
-                var fxInfo = _request.Project.GetTargetFramework(framework);
-                if (fxInfo != null)
-                {
-                    CheckDependencies(result, fxInfo.Dependencies);
-                }
-            }
-
-            return result;
-        }
-
-        private void CheckDependencies(RestoreTargetGraph result, IEnumerable<LibraryDependency> dependencies)
-        {
-            foreach (var dependency in dependencies)
-            {
-                // Ignore floating or version-less (project) dependencies
-                // Avoid warnings for non-packages
-                if (dependency.LibraryRange.TypeConstraintAllows(LibraryDependencyTarget.Package)
-                    && dependency.LibraryRange.VersionRange != null && !dependency.LibraryRange.VersionRange.IsFloating)
-                {
-                    var match = result.Flattened.FirstOrDefault(g => g.Key.Name.Equals(dependency.LibraryRange.Name));
-                    if (match != null 
-                        && LibraryTypes.Package == match.Key.Type
-                        && match.Key.Version > dependency.LibraryRange.VersionRange.MinVersion)
-                    {
-                        _logger.LogWarning(string.Format(CultureInfo.CurrentCulture, Strings.Log_DependencyBumpedUp, 
-                            dependency.LibraryRange.Name,
-                            dependency.LibraryRange.VersionRange.PrettyPrint(),
-                            match.Key.Name,
-                            match.Key.Version));
-                    }
-                }
-            }
-        }
-
-        private Task<RestoreTargetGraph[]> WalkRuntimeDependenciesAsync(LibraryRange projectRange,
-            RestoreTargetGraph graph,
-            IEnumerable<string> runtimeIds,
-            RemoteDependencyWalker walker,
-            RemoteWalkContext context,
-            NuGetv3LocalRepository localRepository,
-            RuntimeGraph runtimes,
-            bool writeToLockFile,
-            CancellationToken token)
-        {
-            var resultGraphs = new List<Task<RestoreTargetGraph>>();
-            foreach (var runtimeName in runtimeIds)
-            {
-                _logger.LogVerbose(string.Format(CultureInfo.CurrentCulture, Strings.Log_RestoringPackages, FrameworkRuntimePair.GetName(graph.Framework, runtimeName)));
-
-                resultGraphs.Add(WalkDependenciesAsync(projectRange,
-                    graph.Framework,
-                    runtimeName,
-                    runtimes,
-                    walker,
-                    context,
-                    writeToLockFile,
-                    token));
-            }
-
-            return Task.WhenAll(resultGraphs);
-        }
-
-        private RuntimeGraph GetRuntimeGraph(RestoreTargetGraph graph, NuGetv3LocalRepository localRepository)
-        {
-            // TODO: Caching!
-            RuntimeGraph runtimeGraph;
-            if (_runtimeGraphCache.TryGetValue(graph.Framework, out runtimeGraph))
-            {
-                return runtimeGraph;
-            }
-
-            _logger.LogVerbose(Strings.Log_ScanningForRuntimeJson);
-            runtimeGraph = RuntimeGraph.Empty;
-            graph.Graphs.ForEach(node =>
-            {
-                var match = node?.Item?.Data?.Match;
-                if (match == null)
-                {
-                    return;
-                }
-
-                // Ignore runtime.json from rejected nodes
-                if (node.Disposition == Disposition.Rejected)
-                {
-                    return;
-                }
-
-                // Locate the package in the local repository
-                var package = localRepository.FindPackagesById(match.Library.Name)
-                    .FirstOrDefault(p => p.Version == match.Library.Version);
-
-                if (package != null)
-                {
-                    var nextGraph = LoadRuntimeGraph(package);
-                    if (nextGraph != null)
-                    {
-                        _logger.LogVerbose(string.Format(CultureInfo.CurrentCulture, Strings.Log_MergingRuntimes, match.Library));
-                        runtimeGraph = RuntimeGraph.Merge(runtimeGraph, nextGraph);
-                    }
-                }
-            });
-            _runtimeGraphCache[graph.Framework] = runtimeGraph;
-            return runtimeGraph;
-        }
-
-        private RuntimeGraph LoadRuntimeGraph(LocalPackageInfo package)
-        {
-            var id = new PackageIdentity(package.Id, package.Version);
-            return _runtimeGraphCacheByPackage.GetOrAdd(id, (x) => LoadRuntimeGraphCore(package));
-        }
-
-        private RuntimeGraph LoadRuntimeGraphCore(LocalPackageInfo package)
-        {
-            var runtimeGraphFile = Path.Combine(package.ExpandedPath, RuntimeGraph.RuntimeGraphFileName);
-            if (File.Exists(runtimeGraphFile))
-            {
-                using (var stream = File.OpenRead(runtimeGraphFile))
-                {
-                    return JsonRuntimeFormat.ReadRuntimeGraph(stream);
-                }
-            }
-            return null;
-        }
-
-        private async Task InstallPackagesAsync(IEnumerable<RestoreTargetGraph> graphs,
-            string packagesDirectory,
-            HashSet<LibraryIdentity> allInstalledPackages,
-            int maxDegreeOfConcurrency,
-            CancellationToken token)
-        {
-            var packagesToInstall = graphs.SelectMany(g => g.Install.Where(match => allInstalledPackages.Add(match.Library)));
-            if (maxDegreeOfConcurrency <= 1)
-            {
-                foreach (var match in packagesToInstall)
-                {
-                    await InstallPackageAsync(match, packagesDirectory, token);
-                }
-            }
-            else
-            {
-                var bag = new ConcurrentBag<RemoteMatch>(packagesToInstall);
-                var tasks = Enumerable.Range(0, maxDegreeOfConcurrency)
-                    .Select(async _ =>
-                        {
-                            RemoteMatch match;
-                            while (bag.TryTake(out match))
-                            {
-                                await InstallPackageAsync(match, packagesDirectory, token);
-                            }
-                        });
-                await Task.WhenAll(tasks);
-            }
-        }
-
-        private async Task InstallPackageAsync(RemoteMatch installItem, string packagesDirectory, CancellationToken token)
-        {
-            var packageIdentity = new PackageIdentity(installItem.Library.Name, installItem.Library.Version);
-
-            var versionFolderPathContext = new VersionFolderPathContext(
-                packageIdentity,
-                packagesDirectory,
-                _logger,
-                fixNuspecIdCasing: true,
-                packageSaveMode: _request.PackageSaveMode,
-                normalizeFileNames: false,
-                xmlDocFileSaveMode: _request.XmlDocFileSaveMode);
-
-            await PackageExtractor.InstallFromSourceAsync(
-                stream => installItem.Provider.CopyToAsync(installItem.Library, stream, token),
-                versionFolderPathContext,
-                token);
         }
     }
 }
